requirements.txt
absl-py==2.1.0
advent-of-code-data==1.3.2
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.15.2
altgraph==0.17.3
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.9.0
bson==0.5.10
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.2.0
contourpy==1.2.0
crispy-bootstrap5==2024.2
cycler==0.12.1
dataclasses-json==0.5.7
DateTime==5.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.1
django-crispy-forms==2.1
engineering-notation==0.8.0
executing==1.2.0
fastapi==0.95.2
fastjsonschema==2.17.1
filelock==3.12.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.5.0
httpx==0.27.0
huggingface==0.0.1
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.28.0
ipython==8.13.2
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.25
jsonpointer==2.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.5
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.1.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.13
jupyterthemes==0.20.0
keras==3.3.3
kiwisolver==1.4.5
langchain==0.0.186
lesscpy==0.15.1
libclang==18.1.1
macholib==1.16.2
Mako==1.3.9
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.3.2
moviepy==2.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.8
networkx==3.1
nodeenv==1.8.0
notebook==7.1.3
notebook_shim==0.2.4
numexpr==2.8.4
numpy==1.26.4
ollama==0.2.0
openai==0.27.6
openapi-schema-pydantic==1.2.4
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.11.0
overrides==7.3.1
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
Pebble==5.0.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.5.1
ply==3.11
proglog==0.1.12
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.1.0
referencing==0.35.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.40
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tk-tools==0.16.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.2.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.9.2
zope.interface==6.4.post2
